refactor : Multimodule Set 멀티모듈을 위한 세팅 gradle 버전업
