+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/Category/CategoryRepository.java
@@ -1,0 +1,6 @@
+package com.example.myecoomerce.myecommercecore.Domain.Category;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
