+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/Comment/CommentRepository.java
@@ -1,0 +1,10 @@
+package com.example.myecoomerce.myecommercecore.Domain.Comment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
+    List<Comment> findCommentsByProductId(Long id);
+}
