+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/ImageFile/ImageFileRepository.java
@@ -1,0 +1,6 @@
+package com.example.myecoomerce.myecommercecore.Domain.ImageFile;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ImageFileRepository extends JpaRepository<ImageFile,Long> {
+}
