+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/Order/OrderRepository.java
@@ -1,0 +1,9 @@
+package com.example.myecoomerce.myecommercecore.Domain.Order;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    public List<Order> findAllByUserUsername(String username);
+}
