+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/Product/ProductRepository.java
@@ -1,0 +1,11 @@
+package com.example.myecoomerce.myecommercecore.Domain.Product;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product,Long > {
+    public List<Product> findAllByUserUsername(String username);
+
+    List<Product> findProductsByCategoryId(Long id);
+}
