+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/User/User.java
@@ -1,0 +1,121 @@
+package com.example.myecoomerce.myecommercecore.Domain.User;
+
+import com.example.myecommerce.Domain.BaseTimeEntity;
+import com.example.myecommerce.Domain.Comment.Comment;
+import com.example.myecommerce.Domain.Order.Order;
+import com.example.myecommerce.Domain.Product.Product;
+import com.example.myecoomerce.myecommercecore.Domain.BaseTimeEntity;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+
+//시큐리티 구현 전까지는 이대로 실습
+@Entity
+@NoArgsConstructor
+@Getter
+@ToString
+public class User extends BaseTimeEntity implements UserDetails {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false,unique = true)
+    private String username;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column
+    private String email;
+
+    @Column
+    private String address;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Role role;
+
+    @OneToMany(mappedBy = "user",cascade = CascadeType.ALL)
+    @JsonManagedReference
+    private Set<Comment> comments = new HashSet<>();
+
+    @OneToMany(mappedBy = "user",cascade = CascadeType.ALL)
+    @JsonManagedReference
+    private Set<Product> products = new HashSet<>();
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+    @JsonManagedReference
+    private Set<Order> orders = new HashSet<>();
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        ArrayList<String> value = new ArrayList();
+        value.add(this.role.getKey());
+        return value.stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList());
+    }
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Builder
+    public User(String username, String password, String name,String email, String address,Role role) {
+        this.username = username;
+        this.password = password;
+        this.name = name;
+        this.email = email;
+        this.address = address;
+        this.role = role;
+    }
+
+    public String getRoleKey(){
+        return this.role.getKey();
+    }
+
+    public void update(String name,String email, String address){
+        this.name = name;
+        this.email = email;
+        this.address = address;
+    }
+
+    public void updatePassword(String password){
+        this.password = password;
+    }
+}
