+++ ecommerce-core/src/main/java/com/example/myecoomerce/myecommercecore/Domain/User/UserRepository.java
@@ -1,0 +1,7 @@
+package com.example.myecoomerce.myecommercecore.Domain.User;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByUsername(String user);
+}
