+++ ecommerce-core/src/main/resources-local/application.properties
@@ -1,0 +1,15 @@
+profile=local
+spring.profiles.include=core
+spring.profiles.active=local
+
+spring.datasource.url=jdbc:mysql://localhost:3306/Ecommerce?serverTimezone=UTC&characterEncoding=UTF-8&useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.username=root
+spring.datasource.password=12341234
+logging.level.org.thymeleaf=info
+spring.mvc.static-path-pattern=/resources/static/**
+
+spring.session.store-type=redis
+server.servlet.session.timeout=10000s
+
+spring.redis.host=127.0.0.1
+spring.redis.port=6379
