+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'my-ecommerce'
+include 'ecommerce-core'
+include 'api'
+
