--- src/main/java/com/example/myecommerce/Domain/Category/CategoryRepository.java
@@ -1,6 +1,0 @@
-package com.example.myecommerce.Domain.Category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
