--- src/main/java/com/example/myecommerce/Domain/Comment/CommentRepository.java
@@ -1,10 +1,0 @@
-package com.example.myecommerce.Domain.Comment;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-
-    List<Comment> findCommentsByProductId(Long id);
-}
