--- src/main/java/com/example/myecommerce/Domain/ImageFile/ImageFileRepository.java
@@ -1,6 +1,0 @@
-package com.example.myecommerce.Domain.ImageFile;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ImageFileRepository extends JpaRepository<ImageFile,Long> {
-}
