--- src/main/java/com/example/myecommerce/Domain/Order/OrderRepository.java
@@ -1,9 +1,0 @@
-package com.example.myecommerce.Domain.Order;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-    public List<Order> findAllByUserUsername(String username);
-}
