--- src/main/java/com/example/myecommerce/Domain/User/Role.java
@@ -1,15 +1,0 @@
-package com.example.myecommerce.Domain.User;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    USER("ROLE_USER","일반사용자"),
-    SELLER("ROLE_SELLER","판매자"),
-    ADMIN("ROLE_ADMIN","관리자");
-
-    private final String key;
-    private final String title;
-}
