--- src/main/java/com/example/myecommerce/Domain/User/UserRepository.java
@@ -1,7 +1,0 @@
-package com.example.myecommerce.Domain.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByUsername(String user);
-}
